requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
fastapi>=0.100.0
uvicorn>=0.22.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
tqdm>=4.65.0
beautifulsoup4>=4.12.0
requests>=2.31.0

# AWS integration
boto3>=1.28.0

# Testing and development
pytest>=7.4.0
black>=23.7.0
flake8>=6.1.0

# API documentation
python-multipart>=0.0.6 